move to literal terminalContrib
