+++ src/vs/workbench/contrib/terminal/terminal.all.ts
@@ -32,3 +32,4 @@ import '../terminalContrib/sendSignal/browser/terminal.sendSignal.contribution.j
 import '../terminalContrib/suggest/browser/terminal.suggest.contribution.js';
 import '../terminalContrib/chat/browser/terminal.initialHint.contribution.js';
 import '../terminalContrib/wslRecommendation/browser/terminal.wslRecommendation.contribution.js';
+import '../terminalContrib/voice/browser/terminal.voice.contribution.js';
