+++ src/vs/workbench/contrib/terminalContrib/voice/browser/terminal.voice.contribution.ts
@@ -1,0 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { registerTerminalVoiceActions } from './terminalVoiceActions.js';
+
+registerTerminalVoiceActions();
